WaterEffectScreen: Refactor to separate wave simulation logic from screen-specific UI, layout, and title rendering.
Modifies:
- TUI.vcxproj
- TUI.vcxproj.filters
-Screens/WaterEffectScreen.h/.cpp

Adds:
- Screens/Content/WaterWaveEffect.h/.cpp

Key changes:

- Introduced WaterWaveEffect (Screens/Content/WaterWaveEffect.h/.cpp)
  - Encapsulates all wave simulation and rendering logic:
    - droplet system and wave propagation
    - rain mode state and timing (Sprinkle / LightRain / Pouring)
    - text mask generation and deformation
    - amplitude calculation and glyph/style selection
  - Renders into a caller-provided viewport (Rect)

- Simplified WaterEffectScreen (Screens/WaterEffectScreen.h/.cpp)
  - Owns all UI/chrome:
    - background fill and frames
    - title/banner animation and fallback handling
    - mode bar display
    - minimum screen warnings
  - Defines viewport via PageComposer and passes it to WaterWaveEffect::draw()

Architectural impact:

- Enables WaterWaveEffect reuse in:
  - windows
  - popups
  - split-screen layouts
  - future widget system
