+++ TUI/Screens/WaterEffectScreen.h
@@ -8,6 +8,7 @@
 #include "Rendering/Objects/LayeredTextObject.h"
 #include "Rendering/Styles/Style.h"
 #include "Rendering/Objects/TextObject.h"
+#include "Screens/Content/WaterWaveEffect.h"
 
 class Surface;
 class ScreenBuffer;
@@ -22,25 +23,6 @@ public:
     void onExit() override;
     void update(double deltaTime) override;
     void draw(Surface& surface) override;
-
-private:
-    struct Droplet
-    {
-        float x = 0.0f;
-        float y = 0.0f;
-        float age = 0.0f;
-        float lifetime = 0.0f;
-        float speed = 0.0f;
-        float wavelength = 0.0f;
-        float strength = 0.0f;
-    };
-
-    enum class RainMode
-    {
-        Sprinkle,
-        LightRain,
-        Pouring
-    };
 
 private:
     // Title Related
@@ -72,19 +54,6 @@ private:
     std::string currentModeTextUtf8() const;
     Style currentModeStyle() const;
 
-    // Simulation Related
-    void ensureSimulationSize(int width, int height);
-    void rebuildTextMask();
-    void spawnRandomDroplet();
-    void updateDroplets(double deltaTime);
-    void updateRainTiming();
-    void chooseNextRainMode();
-    void renderWaveField(Surface& surface);
-    int index(int x, int y) const;
-    int computeAmplitudeAtCell(int x, int y) const;
-    char32_t selectWaveGlyph(int amplitude, char32_t sourceGlyph) const;
-    Style selectWaveStyle(int amplitude, char32_t sourceGlyph) const;
-
 private:
     // Title related
     Assets::AssetLibrary& m_assetLibrary;
@@ -94,6 +63,7 @@ private:
     std::string m_waterTitleLoadError;
     bool m_waterTitleLoadAttempted = false;
     bool m_waterTitleLoaded = false;
+    double m_elapsedSeconds = 0.0;
 
     // Retained static UI cache
     int m_screenWidth = 0;
@@ -105,6 +75,7 @@ private:
     bool m_waterTitleFallbackDirty = true;
 
     RainMode m_cachedModeBarMode = RainMode::Sprinkle;
+    Style m_modeColor[3];
     bool m_cachedWaterTitleLoaded = false;
     std::string m_cachedWaterTitleLoadError;
 
@@ -113,22 +84,5 @@ private:
     TextObject m_minimumScreenUiObject;
     TextObject m_waterTitleFallbackObject;
 
-    // Simulation related
-    int m_waveLeft = 0;
-    int m_waveTop = 0;
-    int m_waveWidth = 0;
-    int m_waveHeight = 0;
-
-    double m_elapsedSeconds = 0.0;
-    double m_nextDropTime = 0.0;
-    double m_nextRainModeChangeTime = 0.0;
-
-    RainMode m_rainMode = RainMode::Sprinkle;
-    Style m_modeColor[3];
-
-    std::u32string m_primaryText = U"WATER EFFECT";
-    std::u32string m_secondaryText = U"THE TEXT BECOMES THE WAVE";
-
-    std::vector<char32_t> m_textMask;
-    std::vector<Droplet> m_droplets;
+    WaterWaveEffect m_waveEffect;
 };
